+++ product/src/main/java/com/kh/product/svc/ProductSVCImpl.java
@@ -28,7 +28,7 @@ public class ProductSVCImpl implements ProductSVC {
     }
 
     @Override
-    public int deleteProduct(Long pid) {
+    public int deleteProduct(int pid) {
         return productDAO.deleteProduct(pid);
     }
